Make the parser a dev dependency where possible
