+++ crates/ruff_python_importer/Cargo.toml
@@ -14,7 +14,6 @@ license = { workspace = true }
 ruff_diagnostics = { workspace = true }
 ruff_python_ast = { workspace = true }
 ruff_python_codegen = { workspace = true }
-ruff_python_parser = { workspace = true }
 ruff_python_trivia = { workspace = true }
 ruff_source_file = { workspace = true, features = ["serde"] }
 ruff_text_size = { workspace = true }
@@ -22,6 +21,8 @@ ruff_text_size = { workspace = true }
 anyhow = { workspace = true }
 
 [dev-dependencies]
+ruff_python_parser = { workspace = true }
+
 insta = { workspace = true }
 
 [features]
