+++ crates/ty_ide/Cargo.toml
@@ -19,7 +19,6 @@ ruff_memory_usage = { workspace = true }
 ruff_python_ast = { workspace = true }
 ruff_python_codegen = { workspace = true }
 ruff_python_importer = { workspace = true }
-ruff_python_parser = { workspace = true }
 ruff_python_trivia = { workspace = true }
 ruff_source_file = { workspace = true }
 ruff_text_size = { workspace = true }
@@ -37,6 +36,8 @@ smallvec = { workspace = true }
 tracing = { workspace = true }
 
 [dev-dependencies]
+ruff_python_parser = { workspace = true }
+
 camino = { workspace = true }
 insta = { workspace = true, features = ["filters"] }
